refactor(partners): migrate partner marketplace into modules/ + hoist shared partner utils
