+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/marketplace/graphql/queries/get-partner-by-slug.ts
@@ -1,17 +1,8 @@
-import { CoreApiClient } from 'twenty-client-sdk/core';
-import { defineLogicFunction } from 'twenty-sdk/define';
-
-import {
-  mapPartnerForMarketplace,
-  type MarketplaceProfilePartner,
-} from './map-partner-for-marketplace';
-
-export const GET_PARTNER_BY_SLUG_LOGIC_FUNCTION_ID =
-  '5e3e7b88-2cf2-4f56-9a4a-46c4c1d6b0bb';
+import { type CoreApiClient } from 'twenty-client-sdk/core';
 
 // CoreApiClient is codegenerated from the synced workspace schema, so the
 // selection is strictly typed and the response shape derives from it.
-const queryPartnerBySlug = (client: CoreApiClient, slug: string) =>
+export const queryPartnerBySlug = (client: CoreApiClient, slug: string) =>
   client.query({
     partners: {
       __args: {
@@ -82,65 +73,3 @@ const queryPartnerBySlug = (client: CoreApiClient, slug: string) =>
       },
     },
   });
-
-type PartnerRaw = NonNullable<
-  Awaited<ReturnType<typeof queryPartnerBySlug>>['partners']
->['edges'][number]['node'];
-
-type GetPartnerBySlugResult =
-  | { ok: true; partner: MarketplaceProfilePartner }
-  | { ok: false; reason: 'NOT_FOUND' | string };
-
-const mapProfilePartner = (node: PartnerRaw): MarketplaceProfilePartner => {
-  const mapped = mapPartnerForMarketplace(node, 'profile');
-
-  if (!('projectBudgetTypical' in mapped)) {
-    throw new Error(
-      'get-partner-by-slug received list payload from profile mapper',
-    );
-  }
-
-  return mapped;
-};
-
-export const handler = async (input: {
-  queryStringParameters?: { slug?: string };
-}): Promise<GetPartnerBySlugResult> => {
-  const slug = input?.queryStringParameters?.slug;
-  if (typeof slug !== 'string' || slug.length === 0) {
-    return { ok: false, reason: 'Missing slug query parameter' };
-  }
-
-  try {
-    const client = new CoreApiClient();
-    const result = await queryPartnerBySlug(client, slug);
-    const rawNode = result.partners?.edges?.[0]?.node;
-
-    if (!rawNode) {
-      return { ok: false, reason: 'NOT_FOUND' };
-    }
-
-    const partner = mapProfilePartner(rawNode);
-
-    return { ok: true, partner };
-  } catch (err) {
-    return {
-      ok: false,
-      reason: err instanceof Error ? err.message : String(err),
-    };
-  }
-};
-
-export default defineLogicFunction({
-  universalIdentifier: GET_PARTNER_BY_SLUG_LOGIC_FUNCTION_ID,
-  name: 'get-partner-by-slug',
-  description:
-    'Returns a single VALIDATED + AVAILABLE partner by slug, or NOT_FOUND.',
-  timeoutSeconds: 10,
-  handler,
-  httpRouteTriggerSettings: {
-    path: '/partner-by-slug',
-    httpMethod: 'GET',
-    isAuthRequired: false,
-  },
-});
