+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/marketplace/list-available-partners.logic-function.ts
@@ -1,0 +1,21 @@
+import { defineLogicFunction } from 'twenty-sdk/define';
+
+import { listAvailablePartners } from 'src/modules/partner/marketplace/services/list-available-partners.service';
+
+export const LIST_AVAILABLE_PARTNERS_LOGIC_FUNCTION_ID =
+  '0f91164f-f492-41e8-9bb0-481be5a3d5b9';
+
+export const handler = listAvailablePartners;
+
+export default defineLogicFunction({
+  universalIdentifier: LIST_AVAILABLE_PARTNERS_LOGIC_FUNCTION_ID,
+  name: 'list-available-partners',
+  description: 'Returns all partners with validationStage=VALIDATED and availability=AVAILABLE.',
+  timeoutSeconds: 10,
+  handler,
+  httpRouteTriggerSettings: {
+    path: '/partners',
+    httpMethod: 'GET',
+    isAuthRequired: false,
+  },
+});
