+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/utils/content-type.ts
