+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/utils/profile-picture.test.ts
